+++ src/plone/restapi/serializer/site.py
@@ -43,7 +43,9 @@ class SerializeSiteRootToJson(object):
             '@id': batch.canonical_url,
             'id': self.context.id,
             '@type': 'Plone Site',
+            'title': self.context.Title(),
             'parent': {},
+            'is_folderish': True
         }
 
         # Insert expandable elements