+++ src/plone/restapi/tests/test_auth.py
@@ -79,6 +79,28 @@ class TestLogin(TestCase):
         res = service.render()
         self.assertIn('token', res)
 
+    def test_login_with_zope_user_fails_without_pas_plugin(self):
+        uf = self.layer['app'].acl_users
+        uf.plugins.users.addUser('zopeuser', 'zopeuser', 'secret')
+        if 'jwt_auth' in uf:
+            uf['jwt_auth'].manage_activateInterfaces([])
+        self.request['BODY'] = '{"login": "zopeuser", "password": "secret"}'
+        service = self.traverse()
+        res = service.reply()
+        self.assertIn('error', res)
+        self.assertEqual('JWT authentication plugin not installed.',
+                         res['error']['message'])
+        self.assertNotIn('token', res)
+
+    def test_login_with_zope_user(self):
+        self.layer['app'].acl_users.plugins.users.addUser(
+            'zopeuser', 'zopeuser', 'secret')
+        self.request['BODY'] = '{"login": "zopeuser", "password": "secret"}'
+        service = self.traverse()
+        res = service.reply()
+        self.assertEqual(200, self.request.response.getStatus())
+        self.assertIn('token', res)
+
 
 class TestLogout(TestCase):