+++ src/plone/restapi/tests/test_documentation.py
@@ -8,20 +8,23 @@ from plone import api
 from plone.app.discussion.interfaces import IConversation
 from plone.app.discussion.interfaces import IDiscussionSettings
 from plone.app.discussion.interfaces import IReplies
-from plone.app.testing import SITE_OWNER_NAME
-from plone.app.testing import SITE_OWNER_PASSWORD
-from plone.app.testing import TEST_USER_ID
+from plone.app.testing import applyProfile
 from plone.app.testing import popGlobalRegistry
 from plone.app.testing import pushGlobalRegistry
 from plone.app.testing import setRoles
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.app.testing import TEST_USER_ID
 from plone.app.textfield.value import RichTextValue
 from plone.locking.interfaces import ITTWLockable
 from plone.namedfile.file import NamedBlobFile
 from plone.namedfile.file import NamedBlobImage
 from plone.registry.interfaces import IRegistry
+from plone.restapi.testing import PAM_INSTALLED
 from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
-from plone.restapi.testing import RelativeSession
+from plone.restapi.testing import PLONE_RESTAPI_DX_PAM_FUNCTIONAL_TESTING
 from plone.restapi.testing import register_static_uuid_utility
+from plone.restapi.testing import RelativeSession
 from plone.testing.z2 import Browser
 from zope.component import createObject
 from zope.component import getUtility
@@ -34,6 +37,10 @@ import re
 import os
 import transaction
 import unittest
+
+if PAM_INSTALLED:
+    from plone.app.multilingual.interfaces import ITranslationManager
+
 
 
 TUS_HEADERS = [
@@ -1225,3 +1232,82 @@ class TestCommenting(unittest.TestCase):
             '/@controlpanels/editing'
         )
         save_request_and_response_for_docs('controlpanels_get_item', response)
+
+
+@unittest.skipUnless(PAM_INSTALLED, 'plone.app.multilingual is installed by default only in Plone 5')
+class TestPAMDocumentation(unittest.TestCase):
+
+    layer = PLONE_RESTAPI_DX_PAM_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer['app']
+        self.request = self.layer['request']
+        self.portal = self.layer['portal']
+        self.portal_url = self.portal.absolute_url()
+
+        self.time_freezer = freeze_time("2016-10-21 19:00:00")
+        self.frozen_time = self.time_freezer.start()
+
+        self.api_session = RelativeSession(self.portal_url)
+        self.api_session.headers.update({'Accept': 'application/json'})
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+
+        language_tool = api.portal.get_tool('portal_languages')
+        language_tool.addSupportedLanguage('en')
+        language_tool.addSupportedLanguage('es')
+        applyProfile(self.portal, 'plone.app.multilingual:default')
+        en_id = self.portal['en'].invokeFactory(
+            'Document',
+            id='test-document',
+            title='Test document'
+        )
+        self.en_content = self.portal['en'].get(en_id)
+        es_id = self.portal['es'].invokeFactory(
+            'Document',
+            id='test-document',
+            title='Test document'
+        )
+        self.es_content = self.portal['es'].get(es_id)
+
+        import transaction
+        transaction.commit()
+        self.browser = Browser(self.app)
+        self.browser.handleErrors = False
+        self.browser.addHeader(
+            'Authorization',
+            'Basic %s:%s' % (SITE_OWNER_NAME, SITE_OWNER_PASSWORD,)
+        )
+
+    def tearDown(self):
+        self.time_freezer.stop()
+
+    def test_documentation_translations_post(self):
+        response = self.api_session.post(
+            '{}/@translations'.format(self.en_content.absolute_url()),
+            json={
+                'id': self.es_content.absolute_url()
+            }
+        )
+        save_request_and_response_for_docs('translations_post', response)
+
+    def test_documentation_translations_get(self):
+        ITranslationManager(self.en_content).register_translation(
+            'es', self.es_content)
+        transaction.commit()
+        response = self.api_session.get(
+            '{}/@translations'.format(self.en_content.absolute_url()))
+
+        save_request_and_response_for_docs('translations_get', response)
+
+    def test_documentation_translations_delete(self):
+        ITranslationManager(self.en_content).register_translation(
+            'es', self.es_content)
+        transaction.commit()
+        response = self.api_session.delete(
+            '{}/@translations'.format(self.en_content.absolute_url()),
+            json={
+                "language": "es"
+            })
+        save_request_and_response_for_docs('translations_delete', response)